% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratter.R
\name{compare_cluster_genes}
\alias{compare_cluster_genes}
\title{Compares Genes Expressed In Each Cluster.}
\usage{
compare_cluster_genes(seurat_object)
}
\arguments{
\item{seurat_object}{A Seurat object that has undergone the entire workflow, involving UMAP.}
}
\value{
Returns a matrix of number of clusters by number of clusters, with each matrix cell (x,y) containing the vector genes expressed by both clusters x and y.
}
\description{
A function that compares and compiles all the genes identified by extract_genes_expressed into one matrix
}
