% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratter.R
\name{determine_dimensionality}
\alias{determine_dimensionality}
\title{Calculates The Optimum Dimensionality From An Elbow Plot.}
\usage{
determine_dimensionality(plot_data)
}
\arguments{
\item{plot}{An elbow plot data used for detemining dimensionality, generated by calling ElbowPlot() on a Seurat object.}
}
\value{
Returns an integer that represents the dimensionality that was determined by calculation.
}
\description{
A function that calculates the optimum dimensionality from an elbow plot. Involves calculating
slopes of the lines that begin from each point to the last point. The optimum dimensionality will
be determined as the point at which slopes no longer change or no visible change occurs to the slopes
thereafter. The plot data must be plotted by dims (x-axis) and stdev (y-axis).
}
