% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratter.R
\name{draw_cluster_gene_relations}
\alias{draw_cluster_gene_relations}
\title{Plot A Graph Representing Gene Expression Relationships Between Clusters}
\usage{
draw_cluster_gene_relations(gene_comparison_matrix)
}
\arguments{
\item{gene_comparison_matrix}{A matrix object of cluster_id by cluster_id cells, with each
matrix cell (x,y) containing the vector genes expressed by both clusters x and y.}
}
\value{
Returns a plot of nodes and edges, where node (number represents cluster id) represents a cluster
and an edge between nodes represent the genes shared by those clusters.
}
\description{
Plot a graph representing the gene expression relationships between clusters. The edge thickness depends
on how many genes the clusters share between each other. The more of the same genes they express, the thicker
stronger their relationship is.
}
