% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elbow_plot_data}
\alias{elbow_plot_data}
\title{Elbow plot data from PBMC}
\format{A list containing 'dims' and 'stdev' data
\describe{
\item{dims}{Dimension number}
\item{stdev}{Standard deviation calculated using number of dimensions in dims.}
}}
\source{
10X Genomics, California, USA
}
\usage{
elbow_plot_data
}
\description{
An elbow plot data generated from PBMC to use in determine_dimensionality().
}
\keyword{datasets}
