% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_comparison_matrix}
\alias{gene_comparison_matrix}
\title{Matrix with cluster id's and genes shared by a pair of clusters.}
\format{A matrix object
\describe{
\item{column}{cluster ids}
\item{row}{cluster ids}
}}
\source{
10X Genomics, California, USA
}
\usage{
gene_comparison_matrix
}
\description{
A matrix with cluster id's as rows and columns.Each cell contains the genes shared by the cluster pair.
}
\keyword{datasets}
