% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PBMC}
\alias{PBMC}
\title{Single Cell RNA Seq Analysis Data}
\format{A Seurat Object (involves many different data that are relevant to the scrna seq workflow but only
the few depicted below are relevant to this pacakge):
\describe{
\item{features}{The number of genes that were expressed by the cells.}
\item{samples}{The number of cells involved in this experiment.}
\item{seurat_clusters}{Identification of clusters for each cell.}
\item{RNA}{Includes the raw and lognormalized gene expression matrices.}
}}
\source{
10X Genomics, California, USA
}
\usage{
PBMC
}
\description{
A Single Cell RNAseq experiment conducted by 10X Genomics on peripheral blood mononuclear cells,
generated by the CellRanger pipeline.
}
\keyword{datasets}
