% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratter.R
\name{plot_cluster_relations}
\alias{plot_cluster_relations}
\title{Plot A Bar Graph Representing Gene Expression Relationships Between Clusters with Numerical Context.}
\usage{
plot_cluster_relations(gene_expression_graph)
}
\arguments{
\item{gene_expression_graph}{Graph output from draw_cluster_relations to utilize the edge weights}
}
\value{
Returns a bar graph indicating which cluster pairs share the more genes with respect to
the pair with the most genes shared.
}
\description{
Plot a bar graph representing the gene expression relationships between clusters. Each bar represents
how much genes are shared in the pair with respect to other pairs. Because the relationships are made
relative to each other, the cluster pair with the most genes shared will be indicated as 1.
}
\examples{
library(igraph)
library(utils)
library(graphics)
gene_comparison_matrix <- compare_cluster_genes(PBMC)
draw_cluster_gene_relations(gene_comparison_matrix)

}
