% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratter.R
\name{run_seuratter}
\alias{run_seuratter}
\title{Combine the functions in seuratter.R to run the workflow for generating the graph.}
\usage{
run_seuratter(seurat_object)
}
\arguments{
\item{seurat_object}{A Seurat object that has undergone the entire workflow, involving UMAP.}
}
\value{
Returns a vector containing all the graphical outputs generated
by seuratter (except determine_dimensionality)
}
\description{
Work as a controller integrating extract_genes_expressed, compare_cluster_genes, and
draw_cluster_gene_relations.
}
\examples{
library(igraph)
library(utils)
library(graphics)
run_seuratter(PBMC)

}
