% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_seuratter.R
\name{run_shiny_seuratter}
\alias{run_shiny_seuratter}
\title{Run the shiny app for seuratter}
\usage{
run_shiny_seuratter()
}
\value{
No return value. Opens a shiny app page.
}
\description{
Run the shiny app for seuratter, displaying all the plots outputted by the package with the example
data PBMC.
}
\examples{
\dontrun{
run_shiny_seuratter()
}

}
