% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratter.R
\name{compare_cluster_genes}
\alias{compare_cluster_genes}
\title{Compares Genes Expressed In Each Cluster.}
\usage{
compare_cluster_genes(seurat_object, min_percent)
}
\arguments{
\item{seurat_object}{A Seurat object that has undergone the entire workflow, involving UMAP.}

\item{min_percent}{A float representing the percentage of cells that have to express the gene to be
included in the result.}
}
\value{
Returns a matrix of number of clusters by number of clusters, with each matrix cell (x,y)
containing the vector of genes expressed by both clusters x and y.
}
\description{
A function that compares and compiles all the genes identified by extract_genes_expressed() into one matrix. Does
this by determining the gene expressions that overlap between clusters and mapping that information into
the matrix.
}
\examples{
compare_cluster_genes(PBMC)

}
