% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratter.R
\name{determine_dimensionality}
\alias{determine_dimensionality}
\title{Calculates The Optimum Dimensionality From An Elbow Plot.}
\usage{
determine_dimensionality(plot_data, cutoff)
}
\arguments{
\item{plot_data}{An elbow plot data used for detemining dimensionality, generated by calling ElbowPlot() on a Seurat object.}

\item{cutoff}{The cutoff that indicates how much data/variance is covered.}
}
\value{
Returns an integer that represents the optimal dimensionality that was determined by calculation.
}
\description{
A function that calculates the optimum dimensionality from an elbow plot. Involves calculating
slopes between two consecutive points. The optimum dimensionality will be determined as the point at
which slopes no longer change or no visible change occurs to the slopes thereafter. In this function, a dimension
is determined optimal when a slope connecting that point to another is flatter than 10 times the flattest slope
in the plot and the stdev at that point is proximal to the lower limit.
The plot data must be plotted by dims (x-axis) and stdev (y-axis).
}
\examples{
library(Seurat)
determine_dimensionality(elbow_plot_data, 0.95)

}
