% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratter.R
\name{determine_dimensionality}
\alias{determine_dimensionality}
\title{Calculate the optimum dimensionality from an elbow plot.}
\usage{
determine_dimensionality(plot)
}
\arguments{
\item{plot}{An elbow plot used for detemining dimensionality.}
}
\value{
This function returns an integer that represents the number of dimensionality calculated.
}
\description{
A function that calculates the optimum dimensionality from an elbow plot. Involves calculating
slopes of the lines that begin from each point to the last point. The optimum dimensionality will
be determined as the point at which slopes no longer change or no visible change occurs to the slopes
thereafter.
}
