% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PBMC}
\alias{PBMC}
\title{Single Cell RNA Seq Analysis Data}
\format{A matrix with columns:
\describe{
\item{Condition.1}{First condition, under which least light was provided
to plants.}
\item{Condition.2}{Second condition, under which most light was provided to plants.}
\item{Condition.3}{Third condition, the control.}
}}
\source{
10X Genomics
}
\usage{
PBMC
}
\description{
A Single Cell RNAseq experiment conducted by 10X Genomics, generated by the CellRanger pipeline.
}
\examples{
\dontrun{
 PBMC
}
}
\keyword{datasets}
